sandra.turner@turner.com.
All statements of fact, opinion, or analysis expressed in this mono- graph are those of the authors. They do not necessarily reflect offi- cial positions or views of the Central Intelligence Agency or any other US Government entity, past or present. Nothing in the con- tents should be construed as asserting or implying US Government endorsement of the study's factual statements and interpretations.  sandra.turner@turner.com.